add safestr type
