+++ draconic/helpers.py
@@ -1,10 +1,10 @@
 import ast
 import operator as op
-from collections import UserList
 
 from .exceptions import *
+from .types import *
 
-__all__ = ("DraconicConfig", "OperatorMixin", "approx_len_of", "safe_dict", "safe_list", "safe_set")
+__all__ = ("DraconicConfig", "OperatorMixin")
 
 # ===== config =====
 DISALLOW_PREFIXES = ['_', 'func_']
@@ -53,6 +53,7 @@ class DraconicConfig:
         self._list = safe_list(self)
         self._dict = safe_dict(self)
         self._set = safe_set(self)
+        self._str = safe_str(self)
 
         # default names
         if default_names is None:
@@ -71,11 +72,15 @@ class DraconicConfig:
     def set(self):
         return self._set
 
+    @property
+    def str(self):
+        return self._str
+
     def _default_names(self):
         return {
             "True": True, "False": False, "None": None,
             # functions
-            "bool": bool, "int": int, "float": float, "str": str, "tuple": tuple,
+            "bool": bool, "int": int, "float": float, "str": self.str, "tuple": tuple,
             "dict": self.dict, "list": self.list, "set": self.set
         }
 
@@ -181,163 +186,3 @@ class OperatorMixin:
             _raise_in_context(NumberTooHigh, "This number is too large")
         if isinstance(b, int) and (b < self._config.min_int or b > self._config.max_int):
             _raise_in_context(NumberTooHigh, "This number is too large")
-
-
-def approx_len_of(obj, visited=None):
-    """Gets the approximate size of an object (including recursive objects)."""
-    if isinstance(obj, str):
-        return len(obj)
-
-    if hasattr(obj, "__approx_len__"):
-        return obj.__approx_len__
-
-    if visited is None:
-        visited = [obj]
-
-    size = op.length_hint(obj)
-
-    if isinstance(obj, dict):
-        obj = obj.items()
-
-    try:
-        for child in iter(obj):
-            if child in visited:
-                continue
-            size += approx_len_of(child, visited)
-            visited.append(child)
-    except TypeError:  # object is not iterable
-        pass
-
-    try:
-        setattr(obj, "__approx_len__", size)
-    except AttributeError:
-        pass
-
-    return size
-
-
-# ===== compound types =====
-# each function is a function that returns a class based on Draconic config
-# ... look, it works
-
-def safe_list(config):
-    class SafeList(UserList):  # extends UserList so that [x] * y returns a SafeList, not a list
-        def __init__(self, *args, **kwargs):
-            super().__init__(*args, **kwargs)
-            self.__approx_len__ = approx_len_of(self)
-
-        def append(self, obj):
-            if approx_len_of(self) + 1 > config.max_const_len:
-                _raise_in_context(IterableTooLong, "This list is too long")
-            super().append(obj)
-            self.__approx_len__ += 1
-
-        def extend(self, iterable):
-            other_len = approx_len_of(iterable)
-            if approx_len_of(self) + other_len > config.max_const_len:
-                _raise_in_context(IterableTooLong, "This list is too long")
-            super().extend(iterable)
-            self.__approx_len__ += other_len
-
-        def pop(self, i=-1):
-            retval = super().pop(i)
-            self.__approx_len__ -= 1
-            return retval
-
-        def remove(self, item):
-            super().remove(item)
-            self.__approx_len__ -= 1
-
-        def clear(self):
-            super().clear()
-            self.__approx_len__ = 0
-
-        def __mul__(self, n):
-            # to prevent the recalculation of the length on list mult we manually set a new instance's
-            # data and approx len (JIRA-54)
-            new = SafeList()
-            new.data = self.data * n
-            new.__approx_len__ = self.__approx_len__ * n
-            return new
-
-    return SafeList
-
-
-def safe_set(config):
-    class SafeSet(set):
-        def __init__(self, *args, **kwargs):
-            super().__init__(*args, **kwargs)
-            self.__approx_len__ = approx_len_of(self)
-
-        def update(self, *s):
-            other_lens = sum(approx_len_of(other) for other in s)
-            if approx_len_of(self) + other_lens > config.max_const_len:
-                _raise_in_context(IterableTooLong, "This set is too large")
-            super().update(*s)
-            self.__approx_len__ += other_lens
-
-        def add(self, element):
-            if approx_len_of(self) + 1 > config.max_const_len:
-                _raise_in_context(IterableTooLong, "This set is too large")
-            super().add(element)
-            self.__approx_len__ += 1
-
-        def union(self, *s):
-            if approx_len_of(self) + sum(approx_len_of(other) for other in s) > config.max_const_len:
-                _raise_in_context(IterableTooLong, "This set is too large")
-            return SafeSet(super().union(*s))
-
-        def pop(self):
-            retval = super().pop()
-            self.__approx_len__ -= 1
-            return retval
-
-        def remove(self, element):
-            super().remove(element)
-            self.__approx_len__ -= 1
-
-        def discard(self, element):
-            super().discard(element)
-            self.__approx_len__ -= 1
-
-        def clear(self):
-            super().clear()
-            self.__approx_len__ = 0
-
-    return SafeSet
-
-
-def safe_dict(config):
-    class SafeDict(dict):
-        def __init__(self, *args, **kwargs):
-            super().__init__(*args, **kwargs)
-            self.__approx_len__ = approx_len_of(self)
-
-        def update(self, other_dict=None, **kvs):
-            if other_dict is None:
-                other_dict = {}
-
-            other_lens = approx_len_of(other_dict) + approx_len_of(kvs)
-            if approx_len_of(self) + other_lens > config.max_const_len:
-                _raise_in_context(IterableTooLong, "This dict is too large")
-
-            super().update(other_dict, **kvs)
-            self.__approx_len__ += other_lens
-
-        def __setitem__(self, key, value):
-            other_len = approx_len_of(value)
-            if approx_len_of(self) + other_len > config.max_const_len:
-                _raise_in_context(IterableTooLong, "This dict is too large")
-            self.__approx_len__ += other_len
-            return super().__setitem__(key, value)
-
-        def pop(self, k):
-            retval = super().pop(k)
-            self.__approx_len__ -= 1
-            return retval
-
-        def __delitem__(self, key):
-            super().__delitem__(key)
-            self.__approx_len__ -= 1
-
-    return SafeDict
