+++ tests/test_limits.py
@@ -141,6 +141,13 @@ def test_types_again(i, e):
     e("b = dict(((1, 1), (2, 2)))")
     assert type(i.names['a']) is type(i.names['b']) is i._dict
 
+    e("a = 'foobar'")
+    e("b = str(123)")
+    assert type(i.names['a']) is type(i.names['b']) is i._str
+
+    i.builtins['typeof'] = lambda o: type(o).__name__
+    assert e('typeof(a)') == 'str'
+
 
 def test_int_limits(e):
     max_int = (2 ** 31) - 1
